Move ZIPL to pyanaconda.bootloader.zipl
Move the ZIPL class and the related code to a new file.
